+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### master
+* Add admin roles message property [#905](https://github.com/synyx/urlaubsverwaltung/pull/905)
 * Remove 'reason' from application logging [#901](https://github.com/synyx/urlaubsverwaltung/pull/901)
 * Remove 'comment' from account logging [#902](https://github.com/synyx/urlaubsverwaltung/pull/902)